Add configuration option to send cell hint to VStream request
@mcrauwel fixed this on airbyte-source in
https://github.com/planetscale/airbyte-source/pull/77, porting the fix
with some of my changes over the vitess connector. The basic issue is
that if edge is connected to a primary's vtgate and we ask to stream
from a replica, vstream will just hang because it doesn't find a
suitable cell to Stream from. Instead, this PR will allow the caller to
give us a cell hint which should help fix this error.

This is what customers would put in their connection properties:

```properties
vitess.cells=planetscale_operator_default
```

Co-authored-by: Phani Raj <PhaniRajuyn@gmail.com>
Signed-off-by: Max Englander <max@planetscale.com>