Emit ddl events for Kafka Connect
This change makes a lot of changes to the existing connector to allow us to
emit DDL events from a PlanetScale database to be consumed by Kafka
Connect.

1. Change `VitessConnectorConfig` to inherit from
   `HistorizedRelationalDatabaseConnectorConfig`  and not
   `RelationalDatabaseConnectorConfig`.
2. Change `VitessDatabaseSchema` to inherit from
   `RelationalDatabaseSchema` to `HistorizedRelationalDatabaseSchema`.
3. Instead of just rotating the offset whenever we encounter a DDL
   event, we will now:
  - Parse the DDL statement using
    [`JSQLParser`](https://jsqlparser.github.io/JSqlParser/) to extract
    the type of operation and the table name.
  - Emit the DDL event to be consumed by whichever kafka consumer is
    listening.

Implement DDL event filtering.

Support TRUNCATE in ddl statements

The existing mysql debezium connector has special handling for the
TRUNCATE DDL statement whenever it appears in the binlog :
https://issues.redhat.com/browse/DBZ-5610

Effectively, it turns the DDL message into a ChangeRecord wtih the
operation set to Truncate.

This change adds something similar for the vitess connector.

Co-authored-by: Phani Raj <PhaniRajuyn@gmail.com>
Signed-off-by: Max Englander <max@planetscale.com>