non-zero tsms fields for snapshot events
The connector currently posts `tsms` equal to `0` for all snapshot
records. This is different from the MySQL connector, which, before
snapshotting each table, does a `SELECT CURRENT_TIMESTAMP`, and uses the
result for `tsms` for each snapshot record produced from rows in that
table.

A customer has told us that the current Vitess connector behavior poses
some problems for them. Namely, they run distinct sets of connectors in
parallel: one set for streaming and the other for snapshots. They can
get a situation where the streaming connector produces an event as
`tsms` equal to `<date>`, and then (after that) the snapshot connector
starts running and produces an event equal to `0`. In this situation
they want the most recent `<date>` to be chosen. The streaming
connector's event will always get picked, but they want the snapshot
connector's event to be preferred because it started after `<date>`.

This change approximates the MySQL connector's behavior. Before the
VStream call we get the `CURRENT_TIMETAMP`, and use the result for each
snapshot record produced from each VStream event until we see
`COPY_COMPLETE` event signaling the transition to streaming.

Signed-off-by: Max Englander <max@planetscale.com>